Small fix
Adding a functionality requirement not mentioned in the instruction document, but the separate project rubric.

Since filenames/paths are currently hardcoded, its been set to output the new .csv to a new file instead of overwriting the original (it was requested to directly overwrite the original.) This can be fixed by setting both filenames to the same thing.